pyproject.toml
[project]
name = "api-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "fastapi>=0.115.8",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.5",
    "markupsafe==2.0.1",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.37",
    "streamlit>=1.42.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.0",
    "websocket-client>=1.8.0",
    "websockets>=14.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4"
]

